requirements.txt
requests
python-dotenv
openai
pydantic

jupyter
ipykernel


instructor
anthropic

streamlit